+++ src/main/java/insane96mcp/enhancedai/modules/base/ai/EASpiderTargetGoal.java
@@ -1,14 +1,13 @@
 package insane96mcp.enhancedai.modules.base.ai;
 
 
-import insane96mcp.insanelib.ai.ILNearestAttackableTargetGoal;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.monster.Spider;
 
 import javax.annotation.Nullable;
 import java.util.function.Predicate;
 
-public class EASpiderTargetGoal<T extends LivingEntity> extends ILNearestAttackableTargetGoal<T> {
+public class EASpiderTargetGoal<T extends LivingEntity> extends EANearestAttackableTarget<T> {
     public EASpiderTargetGoal(Spider goalOwnerIn, Class<T> targetClassIn, boolean checkSight, boolean nearbyOnlyIn, @Nullable Predicate<LivingEntity> targetPredicate) {
         super(goalOwnerIn, targetClassIn, checkSight, nearbyOnlyIn, targetPredicate);
     }