+++ src/main/java/insane96mcp/enhancedai/setup/EAAttributes.java
@@ -10,5 +10,5 @@ import net.minecraftforge.registries.RegistryObject;
 public class EAAttributes {
 	public static final DeferredRegister<Attribute> ATTRIBUTES = DeferredRegister.create(ForgeRegistries.ATTRIBUTES, EnhancedAI.MOD_ID);
 
-	public static final RegistryObject<Attribute> XRAY_FOLLOW_RANGE = ATTRIBUTES.register("generic.xray_follow_range", () -> new RangedAttribute("attribute.name.generic.xray_follow_range", 32d, 0d, 256d));
+	public static final RegistryObject<Attribute> XRAY_FOLLOW_RANGE = ATTRIBUTES.register("generic.xray_follow_range", () -> new RangedAttribute("attribute.name.generic.xray_follow_range", 0d, 0d, 256d));
 }